feat(startshelltask): startShellTask会返回taskId，在日志读取失败（或为空）时不会删除日志文本

+++ app/extend/context.js
@@ -3,7 +3,7 @@ module.exports = {
   /**
    * 启动新线程执行异步shell任务
    * @param shellTaskFilename {string} shell任务文件名，文件名应是app/shell/目录下的文件，不包含路径名
-   * @returns {Promise<void>}
+   * @returns {Promise<string>} 返回taskId
    */
   async startShellTask(shellTaskFilename) {
     const fs = require('fs');
@@ -24,12 +24,16 @@ module.exports = {
     });
 
     const writerStream = fs.createWriteStream(shellOutputPath);
+    this.logger.info(
+      `[${taskId}] start shell task: ${shellTaskFilename},log path: ${shellOutputPath}`
+    );
     const { fork } = require('child_process');
+
     const forked = fork(`app/shell/${shellTaskFilename}.js`, [taskId], {
       silent: true,
     });
     forked.on('message', async ({ msg }) => {
-      this.logger.error(`{${taskId}} 执行shell子线程时出现异常：${msg}`, msg);
+      this.logger.error(`{${taskId}} 执行shell子线程时出现异常：${msg}`);
       writerStream.write(
         `[${dayjs().format('YYYY-MM-DD HH:mm:SSS')}] ${msg}`,
         'utf8'
@@ -49,23 +53,29 @@ module.exports = {
       );
     });
     forked.stdout.on('end', async () => {
-      writerStream.write(
+      writerStream.end(
         `[${dayjs().format('YYYY-MM-DD HH:mm:SSS')}] Shell Finished`,
         'utf8'
       );
-      writerStream.end();
       this.logger.info(`[${taskId}] shell 任务结束`);
 
-      await task.update({
-        state: isFailed ? FAILED : FINISHED,
-        log: fs.readFileSync(shellOutputPath, 'utf8'),
-      });
-      this.logger.info(`[${taskId}] 日志已写入数据库并更新任务状态`);
+      fs.readFile(shellOutputPath, { encoding: 'utf8' }, async (err, data) => {
+        await task.update({
+          state: isFailed ? FAILED : FINISHED,
+          log: data,
+        });
+        this.logger.info(`[${taskId}] 日志已写入数据库并更新任务状态`);
 
-      fs.unlinkSync(shellOutputPath);
-      this.logger.info(`[${taskId}] 日志文件已删除`);
+        if (err || !data) {
+          this.logger.error(`[${taskId}] 日志读取失败，不执行日志文件删除操作`);
+        } else {
+          fs.unlinkSync(shellOutputPath);
+          this.logger.info(`[${taskId}] 日志文件已删除`);
+        }
+      });
     });
 
+    return taskId;
     /**
      * 获取shell任务名称
      * @param shellTaskFilename  {string} shell任务文件名，文件名应是app/shell/目录下的文件，不包含路径名
@@ -73,8 +83,6 @@ module.exports = {
      */
     function getTaskName(shellTaskFilename) {
       switch (shellTaskFilename) {
-        case 'initGitRepository':
-          return '初始化git仓库';
         case 'initRecoXTemplate':
           return '使用模板VuePressTemplate-recoX初始化VuePress';
         default:
